chore: rename files pkg to edit
