+++ pkg/edit/editor.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package files
+package edit
 
 import (
 	"context"
